+++ Changes.md
@@ -1,3 +1,4 @@
+* #1331 - Add editing of presenters for virtual subjective categories
 * #1327 - Fix problem with inconsistent CSV writing and reading
 
 Release 20.3.0